refactor(ui/plain-text-viewer): Move apiUtils into utils folder to reduce content at project root
