+++ pages/api/dataset/[datasetId]/[datasetVersion].ts
@@ -1,7 +1,7 @@
 import { withApiAuthRequired } from '@auth0/nextjs-auth0';
 
-import { handleApiError } from '../../../../apiUtils/handleApiError';
-import { restreamDownload } from '../../../../apiUtils/restreamDownload';
+import { handleApiError } from '../../../../utils/api/handleApiError';
+import { restreamDownload } from '../../../../utils/api/restreamDownload';
 
 export default withApiAuthRequired(async (req, res) => {
   const { datasetId, datasetVersion } = req.query;
