+++ utils/api/FixedSizeStreamTransform.ts
